[Gold III] Title: 최소비용 구하기 2, Time: 364 ms, Memory: 53896 KB -BaekjoonHub

+++ 백준/Gold/11779. 최소비용 구하기 2/최소비용 구하기 2.java
@@ -3,19 +3,34 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.*;
 
+/**
+ * 다익스트라 문제
+ */
 public class Main {
+    /**
+     * 3 가지 용도
+     *  1. 출발 도시부터 도착 도시까지 비용 정보 -> weights 에 저장
+     *      도착 도시
+     *      이동 비용
+     *  2. 시작 도시부터 현재 도시까지 총합 비용 정보 -> pq 에 저장
+     *      다음 도착 도시
+     *      다음 도시까지 이동 총 합 비용
+     *  3. 현재 도시까지 가장 낮은 비용 정보 -> confirmed 에 저장
+     *      현재 도시에 도착하기 이전 도시
+     *      현재 도시까지 이동 총 합 비용
+     */
     static class Info implements Comparable<Info>{
         int city;
-        long weight;
+        int weight;
 
-        public Info(int city, long weight) {
+        public Info(int city, int weight) {
             this.city = city;
             this.weight = weight;
         }
 
         @Override
         public int compareTo(Info o) {
-            return Long.compare(weight, o.weight);
+            return Integer.compare(weight, o.weight);
         }
     }
 
@@ -25,6 +40,8 @@ public class Main {
         int n = Integer.parseInt(br.readLine());
         int m = Integer.parseInt(br.readLine());
 
+        // 도시 간 이동 비용 정보 저장
+        // 비용이 0인 경우도 있기에 배열 대신 list 로 진행
         List<Info>[] weights = new ArrayList[n + 1];
         for (int i = 1; i <= n; i++)
             weights[i] = new ArrayList();
@@ -33,7 +50,7 @@ public class Main {
             StringTokenizer st = new StringTokenizer(br.readLine());
             int start = Integer.parseInt(st.nextToken());
             int end = Integer.parseInt(st.nextToken());
-            long weight = Long.parseLong(st.nextToken());
+            int weight = Integer.parseInt(st.nextToken());
 
             weights[start].add(new Info(end, weight));
         }
@@ -61,7 +78,7 @@ public class Main {
                 break;
 
             for (Info next : weights[now.city]) {
-                long newWeight = now.weight + next.weight;
+                int newWeight = now.weight + next.weight;
                 if (confirmed[next.city] != null && confirmed[next.city].weight <= newWeight)
                     continue;
